requirements.txt
# requirements.txt (actualizado)
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
gunicorn==21.2.0

# Testing
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0  # Para medir cobertura
factory-boy==3.3.0  # Para crear datos de prueba fácilmente
faker==19.3.0  # Para generar datos falsos